+++ src/Http/Requests/UpdateCustomer.php
@@ -17,6 +17,7 @@ namespace Konekt\AppShell\Http\Requests;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 use Konekt\AppShell\Contracts\Requests\UpdateCustomer as UpdateCustomerContract;
+use Konekt\AppShell\Settings\DefaultCurrency;
 use Konekt\Customer\Models\CustomerTypeProxy;
 
 class UpdateCustomer extends FormRequest implements UpdateCustomerContract
@@ -33,13 +34,23 @@ class UpdateCustomer extends FormRequest implements UpdateCustomerContract
             'firstname' => 'required_if:type,individual',
             'lastname' => 'required_if:type,individual',
             'company_name' => 'required_if:type,organization',
-            'is_active' => 'sometimes|boolean'
+            'is_active' => 'sometimes|boolean',
+            'timezone' => 'nullable|timezone',
+            'ltv' => 'nullable|numeric',
+            'currency' => ['nullable', Rule::in(array_keys((new DefaultCurrency())->options()))],
         ];
     }
 
-    /**
-     * @inheritDoc
-     */
+    public function all($keys = null)
+    {
+        $result = parent::all($keys);
+        if (null === $result['ltv']) {
+            $result['ltv'] = 0;
+        }
+
+        return $result;
+    }
+
     public function authorize()
     {
         return true;