+++ src/resources/views/customer/edit.blade.php
@@ -7,16 +7,30 @@
 @section('content')
 
     {!! Form::model($customer, ['route' => ['appshell.customer.update', $customer], 'method' => 'PUT']) !!}
-    @component(theme_widget('group'), ['accent' => 'secondary'])
-        @slot('title'){{ __('Customer Details') }}@endslot
+        <div class="row">
 
-        @include('appshell::customer._form')
+            <div class="col-md-8 col-lg-9">
+                @component(theme_widget('group'), ['accent' => 'secondary'])
+                    @slot('title'){{ __('Customer Details') }}@endslot
 
-        @slot('footer')
+                    @include('appshell::customer._form')
+                @endcomponent
+            </div>
+
+            <div class="col-md-4 col-lg-3">
+                @component(theme_widget('group'), ['accent' => 'secondary'])
+                    @slot('title'){{ __('Settings') }}@endslot
+
+                    @include('appshell::customer._settings')
+                @endcomponent
+            </div>
+
+        </div>
+
+        @component(theme_widget('group'))
             <button class="btn btn-primary">{{ __('Save') }}</button>
             <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
-        @endslot
-    @endcomponent
+        @endcomponent
     {!! Form::close() !!}
 
 @stop